+++ src/main/java/view/InputView.java
@@ -8,7 +8,6 @@ import java.util.Scanner;
 import static java.util.stream.Collectors.toList;
 
 public class InputView {
-    private static final String YES = "y";
     private static final String SEPARATOR = ",";
 
     private static Scanner scanner = new Scanner(System.in);
@@ -22,8 +21,8 @@ public class InputView {
         return new BigDecimal(scanner.next());
     }
 
-    public static boolean inputAnswer() {
-        return YES.equals(scanner.next().toLowerCase());
+    public static String inputAnswer() {
+        return scanner.next().toLowerCase();
     }
 
 }
