+++ src/main/java/view/OutputView.java
@@ -1,7 +1,6 @@
 package view;
 
 import domain.card.Card;
-import domain.game.BlackJackGame;
 import domain.result.PlayerProfitStatistics;
 import domain.user.*;
 
@@ -19,9 +18,7 @@ public class OutputView {
         System.out.println(playerName.getName() + "의 배팅 금액은?");
     }
 
-    public static void printDistributeCardsMessageAndCardsOfAllUsers(BlackJackGame blackJackGame) {
-        Players players = blackJackGame.getPlayers();
-        Dealer dealer = blackJackGame.getDealer();
+    public static void printDistributeCardsMessageAndCardsOfAllUsers(Players players, Dealer dealer) {
         printHandedOutTwoCardsToPlayers(players);
         printCardsOfAllUsers(players, dealer);
     }
@@ -76,9 +73,7 @@ public class OutputView {
         System.out.println("딜러는 16이하라 한장의 카드를 더 받았습니다.");
     }
 
-    public static void printDealerAndPlayersResult(BlackJackGame blackJackGame) {
-        Players players = blackJackGame.getPlayers();
-        Dealer dealer = blackJackGame.getDealer();
+    public static void printDealerAndPlayersResult(Players players, Dealer dealer) {
         printDealerResult(dealer);
         players.getPlayers()
                 .forEach(OutputView::printPlayerResult);
@@ -98,9 +93,7 @@ public class OutputView {
         System.out.println(playerResultFormat);
     }
 
-    public static void printDealerAndPlayersProfit(BlackJackGame blackJackGame) {
-        Players players = blackJackGame.getPlayers();
-        Dealer dealer = blackJackGame.getDealer();
+    public static void printDealerAndPlayersProfit(Players players, Dealer dealer) {
         PlayerProfitStatistics playerProfitStatistics = players.producePlayersProfitStatistics(dealer);
         printDealerProfit(playerProfitStatistics);
         printPlayersProfit(players, playerProfitStatistics);
